# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Public API documentation served through QRUPI API Gateway.",
        "title": "QRUPI V2 API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/learning-resources/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Learning Resource API"
                ],
                "summary": "Get Learning Resource by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Learning Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Learning Resource API"
                ],
                "summary": "Update Learning Resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Learning Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "file_url",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "learning_group_ids",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "file",
                            "video",
                            "link"
                        ],
                        "type": "string",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "uploaded_user_id",
                        "in": "formData"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Learning Resource API"
                ],
                "summary": "Delete Learning Resource permanently",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Learning Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}